Move color functions to own file
